+++ EnergyProducerService/src/main/java/com/Distr_Sys/producer/repository/ProductionRepository.java
@@ -2,11 +2,8 @@ package com.Distr_Sys.producer.repository;
 
 import com.Distr_Sys.producer.model.ProductionRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface ProductionRepository extends JpaRepository<ProductionRecord, Integer> {
+public interface ProductionRepository extends JpaRepository<ProductionRecord, Long> {
     Optional<ProductionRecord> findTopByOrderByDatetimeDesc();
-
 }
